Move upgrader functions to sqlite.db_upgrades
